requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.3.1
billiard==3.3.0.23
boto==2.38.0
celery==3.1.23
chardet==2.3.0
conceptpower-api
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.6
django-autocomplete-light==3.1.5
django-braces==1.8.1
django-celery==3.1.17
django-cors-headers==1.1.0
django-extensions==1.6.3
django-filter==0.10.0
django-guardian==1.4.4
-e git://github.com/erickpeirson/django-haystack.git@d42090f5dda4323e5d377a9ec4fda719c8ea365d#egg=django_haystack-origin_es2
django-oauth-plus==2.2.8
django-oauth-toolkit==0.9.0
django-oauth2==1.1
django-oauth2-provider==0.2.6.1
django-querysetsequence==0.5
django-storages==1.1.8
djangorestframework==3.3.3
djangorestframework-oauth==1.0.1
elasticsearch==2.3.0
gunicorn==19.3.0
html5lib==0.999999
httplib2==0.9.1
iso8601==0.1.10
isodate==0.5.1
kombu==3.0.35
Markdown==2.6.2
oauth2==1.9.0.post1
oauth2-provider==0.0
oauthlib==1.0.1
pdfminer==20110515
pyparsing==2.0.3
python-magic==0.4.6
pytz==2016.4
rdflib==4.2.0
redis==2.10.5
requests==2.7.0
shortuuid==0.4.2
six==1.10.0
slate==0.3
SPARQLWrapper==1.6.4
static3==0.6.1
Unidecode==0.4.18
urllib3==1.15.1
wheel==0.24.0
whitenoise==2.0.4
jsonpickle
